refactorings, reduced duplication, DFViewer now is in terms of Dfviewerinfinite
